+++ recipes/setup_kvm.rb
@@ -15,10 +15,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-abiquo_nfs node['abiquo']['nfs']['mountpoint'] do
-    share node['abiquo']['nfs']['location']
-    oldshare "10.60.1.72:/opt/vm_repository"
-    action :configure
+mount node['abiquo']['nfs']['mountpoint'] do
+    device node['abiquo']['nfs']['location']
+    fstype "nfs"
+    action [:enable, :mount]
     not_if { node['abiquo']['nfs']['location'].nil? }
 end